hack: temporarily use next.config.js instead of next.config.ts (#5805)
